2025 android

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Goo.java
@@ -22,6 +22,7 @@
 package com.shatteredpixel.shatteredpixeldungeon.actors.mobs;
 
 import static com.shatteredpixel.shatteredpixeldungeon.Dungeon.hero;
+import static com.shatteredpixel.shatteredpixeldungeon.Dungeon.level;
 
 import com.shatteredpixel.shatteredpixeldungeon.Badges;
 import com.shatteredpixel.shatteredpixeldungeon.Challenges;
@@ -37,6 +38,8 @@ import com.shatteredpixel.shatteredpixeldungeon.effects.Speck;
 import com.shatteredpixel.shatteredpixeldungeon.items.Generator;
 import com.shatteredpixel.shatteredpixeldungeon.items.artifacts.DriedRose;
 import com.shatteredpixel.shatteredpixeldungeon.items.keys.CrystalKey;
+import com.shatteredpixel.shatteredpixeldungeon.items.keys.SkeletonKey;
+import com.shatteredpixel.shatteredpixeldungeon.levels.Terrain;
 import com.shatteredpixel.shatteredpixeldungeon.mechanics.Ballistica;
 import com.shatteredpixel.shatteredpixeldungeon.messages.Messages;
 import com.shatteredpixel.shatteredpixeldungeon.scenes.GameScene;
@@ -45,6 +48,7 @@ import com.shatteredpixel.shatteredpixeldungeon.sprites.CharSprite;
 import com.shatteredpixel.shatteredpixeldungeon.sprites.GooSprite;
 import com.shatteredpixel.shatteredpixeldungeon.ui.BossHealthBar;
 import com.shatteredpixel.shatteredpixeldungeon.utils.GLog;
+import com.watabou.noosa.Camera;
 import com.watabou.utils.Bundle;
 import com.watabou.utils.GameMath;
 import com.watabou.utils.PathFinder;
@@ -292,8 +296,10 @@ public class Goo extends Mob {
 
 		//Dungeon.level.drop( Generator.random( Generator.Category.WAND), pos) .sprite.drop();
 
-		Dungeon.level.drop( new CrystalKey( Dungeon.depth ), hero.pos ).sprite.drop();
-		Dungeon.level.drop( new CrystalKey( Dungeon.depth ), pos ).sprite.drop();
+		if(Dungeon.branch == 2) {
+			Dungeon.level.drop(new CrystalKey(Dungeon.depth), hero.pos).sprite.drop();
+			Dungeon.level.drop(new CrystalKey(Dungeon.depth), pos).sprite.drop();
+		}
 
 		//60% chance of 2 blobs, 30% chance of 3, 10% chance for 4. Average of 2.5
 		int blobs = Random.chances(new float[]{0, 0, 6, 3, 1});
@@ -301,11 +307,21 @@ public class Goo extends Mob {
 			int ofs;
 			do {
 				ofs = PathFinder.NEIGHBOURS8[Random.Int(8)];
-			} while (!Dungeon.level.passable[pos + ofs]);
-			Dungeon.level.drop( Generator.random( Generator.Category.STONE), pos + ofs ).sprite.drop( pos );
+			} while (!Dungeon.level.passable[level.map[pos] == Terrain.CHASM ? level.entrance() : pos + ofs]);
+			if(level.map[pos] == Terrain.CHASM){
+				Dungeon.level.drop( Generator.random( Generator.Category.STONE), level.entrance() + ofs ).sprite.drop();
+			} else {
+				Dungeon.level.drop( Generator.random( Generator.Category.STONE), pos + ofs ).sprite.drop( pos );
+			}
 		}
 
-
+		if(Statistics.bossRushMode){
+			if(level.map[pos] == Terrain.CHASM){
+				Dungeon.level.drop( new SkeletonKey( Dungeon.depth ), level.entrance() ).sprite.drop();
+			} else {
+				Dungeon.level.drop( new SkeletonKey( Dungeon.depth ), pos ).sprite.drop();
+			}
+		}
 
 
 		Badges.GOO();
@@ -320,6 +336,8 @@ public class Goo extends Mob {
 		if (!BossHealthBar.isAssigned()) {
 			BossHealthBar.assignBoss(this);
 			Dungeon.level.seal();
+			Camera.main.shake(1f,3f);
+			GameScene.bossReady();
 			yell(Messages.get(this, "notice"));
 			for (Char ch : Actor.chars()){
 				if (ch instanceof DriedRose.GhostHero){
